backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
openai
psycopg2-binary
tiktoken
faiss-cpu
langchain-openai
langchain-community
